pyproject.toml
[tool.poetry]
name = "mariocesar.xyz"
version = "0.1.0"
description = ""
authors = ["Mario César <mariocesar.c50@gmail.com>"]
packages = [{ include = "builder", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
Jinja2 = "^3.1.1"
PyYAML = "^6.0"
yattag = "^1.14.0"
Markdown = "^3.3.6"
mdx-truly-sane-lists = "^1.2"
click = "^8.1.2"
watchdog = "^2.1.7"
fastapi = "^0.78.0"
uvicorn = {extras = ["uvloop"], version = "^0.17.6"}

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
black = "^22.3.0"
flake8 = "^4.0.1"
flake8-bugbear = "^22.3.23"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 100
profile = "black"
default_section = "THIRDPARTY"
known_first_party = "app"

[tool.black]
target-version = ['py39']
line-length = 100
